Revamp expression handling to detect await in func/arrow args
This was an unfortunate case of unforseen design decisions. The last test in test262 was about detecting await inside an (arbitrary) arrow arg default, basically this case: `async (foo = (await bar)) => {}`. To be able to detect this without having to backtrack I have to use the return value of pretty much all the expression parsers but those were already taken by messaging about assignability. So I made that a bit-field to piggyback on it.

This whole thing is only half of it because I need to give `yield` pretty much the same treatment. Although that diff should be much smaller.

This also fixes a case where `await` was incorrectly using strict mode to determine keyword status (it's module goal based, if nothing else).

A bunch of small things were improved in the process. The code was made vastly uglier in the process. Expect cleanup "soon".