Implement line tracking
This will add line/col information to the AST where appropriate.

The majority of the commit updates all usages of `AST_open` to receive the token that opens each individual AST node. This way the start node is always set properly. And since it turns out that all `AST_close` cases can just refer to curtok, no further changes were required for the end.

Note that I only spot checked a few cases. We'll have to fix forward whatever is broken in that department. We can compare to the output of other parsers and fix if we see a discrepancy.