Fix some unicode cases parsing named group idents in regular expressions
It also added a slew of new test cases to cover all edge cases, including unicode escaped surrogate pairs and ID_Start and ID_Continue mixed cases.

Also improves the error being thrown for various regex errors. Before they tended to be squashed by generic messages in the end. Now it will report the first error that was seen. (Still not ideal and I'll probably circle back to it later)

Unfortunately it does not fix the case I tried to fix, but it certainly improved the situation :)