Fix a bunch of regex unicode edge cases, encode unicode chars in test cases
There was a problem where partial surrogate pairs were not properly written to test files. This led to test cases failing unnoticed. The problem is, in part, due to how utf is written. It refuses to write one part of a surrogate pair, which means I can't write tests that use those chars as literals.

As a solution I'm now simply encoding anything out of printable ascii range as `@{x123}@` and decoding/encoding that transparently on the file-io level.

Also solved some small bugs lingering in the unicode handling in char classes in regexes.

Added the `--force-write` option to the test runner to write the resulting output of a test regardless of whether a change was detected (needed this to work around the surrogate pair problem);