Revamp the token type composition (huge change)
Before the token type was a simple bitfield where every bit stood for a particular type. Some bits would be a super type, like number vs dec/hex/oct/bin or string or template, others augmented the token type, like bad escapes for templates.

The problem is that in JS, bitwise operations are always 32bit. That is a hard limit. And we are running out of space.

To work around this limit the token type is now partially bitfield and partially a continuous space. By reserving the first n bits as a continuous space, we can define 32 mutually exclusive types with just 5 bits (a decimal number is never a double quoted string).

Refactoring the whole thing freed up 17 bits because we now use 10 bits for high level token type group flags, like number or template, and 5 bits to create a range from 0 to 31 of continuous space for the other types.

This was necessary because I needed to add a flag for Bigint and I ran out of space :)

With the new available space we can also consider to do the same trick for specific token types, potentially saving some (re)parse time by passing on informtion the lexer already knew (like exactly which token was parsed, etc). This could save some "expensive" string comparisons and replace them with cheaper int comparisons. yay!