Fold up expr/stmt start lexer abstractions by duping the expr lexer (tbc)
There is some regression in error reporting quality since the lexer supersedes the parser errors so some semantic errors are currently lost. Not a prio for me right now.

Since those lexers are now exhaustive, they don't have to be wrapped with the fallback. That should be more efficient... should be. Will probably do this for every lexer soon-if-not-immediately.